feat(api): share profileReward with organization when rewarding user for combined course
